+++ src/modules/robot/arm_solutions/BaseSolution.h
@@ -6,14 +6,14 @@ class BaseSolution {
     public:
         BaseSolution();
         virtual ~BaseSolution() {};
-        virtual void millimeters_to_steps( double millimeters[], int steps[] ) = 0;
-        virtual void steps_to_millimeters( int steps[], double millimeters[] ) = 0;
+        virtual void millimeters_to_steps( float millimeters[], int steps[] ) = 0;
+        virtual void steps_to_millimeters( int steps[], float millimeters[] ) = 0;
 
-        virtual void set_steps_per_millimeter( double steps[] ) = 0;
-        virtual void get_steps_per_millimeter( double steps[] ) = 0;
+        virtual void set_steps_per_millimeter( float steps[] ) = 0;
+        virtual void get_steps_per_millimeter( float steps[] ) = 0;
 
-        virtual bool set_optional(char parameter, double value) { return false; };
-        virtual bool get_optional(char parameter, double *value) { return false; };
+        virtual bool set_optional(char parameter, float value) { return false; };
+        virtual bool get_optional(char parameter, float *value) { return false; };
 };
 
 #endif
