+++ src/modules/robot/arm_solutions/RostockSolution.h
@@ -21,11 +21,11 @@
 class RostockSolution : public BaseSolution {
     public:
         RostockSolution(Config* passed_config);
-        void millimeters_to_steps( double millimeters[], int steps[] );
-        void steps_to_millimeters( int steps[], double millimeters[] ); 
+        void millimeters_to_steps( float millimeters[], int steps[] );
+        void steps_to_millimeters( int steps[], float millimeters[] ); 
 
-        void set_steps_per_millimeter( double steps[] );
-        void get_steps_per_millimeter( double steps[] );
+        void set_steps_per_millimeter( float steps[] );
+        void get_steps_per_millimeter( float steps[] );
 
         float solve_arm( float millimeters[] );
         void rotate( float in[], float out[], float sin, float cos );
